+++ src/Components/HomeMain/HomeMain.module.css
@@ -1,0 +1,125 @@
+.main_container {
+    display: flex;
+    flex-direction: column;
+    row-gap: 1rem;
+    align-items: center;
+    height: fit-content;
+    padding: 1em 0;
+}
+.main_container h2 {
+    font-size: 30px;
+    position: relative;
+    color: var(--white);
+    font-family: var(--font-family-grotesk);
+}
+.main_container > p {
+    padding-bottom: 0.5em;
+    position: relative;
+    font-family: var(--font-family-grotesk);
+    color: var(--white-smoke);
+}
+.main_container > p::after {
+    content: "";
+    background: var(--main-color);
+    height: 1px;
+    width: 50px;
+    position: absolute;
+    bottom: 0;
+    left: 50%;
+    transform: translateX(-50%);
+}
+.main_content {
+    display: flex;
+    align-items: start;
+    justify-content: space-between;
+    column-gap: 1em;
+    width: 100%;
+    height: 100%;
+    position: relative;
+}
+.main_content > .main_image {
+    width: 400px;
+    & img {
+        width: 100%;
+        opacity: 1;
+        height: 100%;
+        transition: all 1s;
+    }
+}
+.fixed {
+    position: sticky;
+    top: 80px;
+    right: 0;
+}
+.absolute {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+}
+
+.reasons {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 1em;
+    width: 600px;
+}
+
+.reason {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: start;
+    row-gap: 1em;
+    padding: 1em;
+    width: 300px;
+    border-radius: 12px;
+    background-color: transparent;
+    background-color: #222;
+    box-shadow: rgba(0, 0, 0, 0.4) 0px 2px 4px,
+        rgba(0, 0, 0, 0.3) 0px 7px 13px -3px,
+        rgba(0, 0, 0, 0.2) 0px -3px 0px inset;
+}
+.reasons > .reason:nth-child(9) {
+    background-color: var(--main-color);
+    grid-column: span 2;
+    grid-row: span 2;
+    width: 100%;
+    & .reason_top h2 {
+        border: 1px solid var(--white);
+    }
+}
+.reason_top {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    color: var(--main-color);
+}
+.reason_top h2 {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 1px solid var(--main-color);
+    color: var(--white);
+    height: 50px;
+    width: 50px;
+    border-radius: 50%;
+}
+.reason_bottom {
+    padding: 0.5em;
+    display: flex;
+    flex-direction: column;
+    row-gap: 0.5em;
+}
+.reason_bottom h3 {
+    font-family: var(--font-family-grotesk);
+    font-size: 16px;
+    font-weight: 500;
+    color: var(--white);
+}
+.reason_bottom p {
+    font-family: var(--font-family-grotesk);
+    font-size: 13px;
+    color: var(--white-smoke);
+    font-weight: 500;
+}
