+++ src/Components/HomeSection/HomeSection.module.css
@@ -1,24 +1,20 @@
-.main {
+.section_cont {
     display: flex;
-    flex-direction: column;
-    row-gap: 1rem;
+    width: 100%;
+    justify-content: space-between;
     align-items: center;
-    height: fit-content;
-    padding: 1em 0;
+    justify-content: center;
+    flex-direction: column;
+    row-gap: 3rem;
+    overflow: hidden;
+    & h2 {
+        color: var(--white-smoke);
+        font-family: var(--font-family-grotesk);
+        position: relative;
+        padding: 0.5rem;
+    }
 }
-.main h2 {
-    font-size: 30px;
-    position: relative;
-    color: var(--white);
-    font-family: var(--font-family-grotesk);
-}
-.main > p {
-    padding-bottom: 0.5em;
-    position: relative;
-    font-family: var(--font-family-grotesk);
-    color: var(--white-smoke);
-}
-.main > p::after {
+.section_cont h2::after {
     content: "";
     background: var(--main-color);
     height: 1px;
@@ -28,60 +24,58 @@
     left: 50%;
     transform: translateX(-50%);
 }
-.reasons {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 1em;
+.section {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+    column-gap: 2rem;
 }
-
-.reason {
+.section_text {
     display: flex;
     flex-direction: column;
+    align-items: start;
+    text-align: start;
+    width: 100%;
+    & p {
+        color: var(--white-smoke);
+        font-family: var(--font-family-grotesk);
+        text-align: start;
+        width: 400px;
+        font-size: 20px;
+    }
+}
+
+.section_video_cont {
+    display: flex;
     align-items: center;
-    justify-content: start;
-    row-gap: 1em;
-    padding: 1em;
-    width: 300px;
-    border-radius: 12px;
+    justify-content: center;
+    object-position: center;
+    padding: 1rem;
+    position: relative;
     background-color: transparent;
-    background-color: #222;
-    box-shadow: rgba(0, 0, 0, 0.4) 0px 2px 4px,
-        rgba(0, 0, 0, 0.3) 0px 7px 13px -3px,
-        rgba(0, 0, 0, 0.2) 0px -3px 0px inset;
+    height: 300px;
+    width: 100%;
+    border-radius: 20px;
+}
+.section_video_cont video {
+    border-radius: 25px;
+    clip-path: fill-box;
+    height: 100%;
 }
 
-.reason_top {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    color: var(--main-color);
-}
-.reason_top h2 {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 1px solid var(--main-color);
-    color: var(--white);
-    height: 50px;
-    width: 50px;
-    border-radius: 50%;
-}
-.reason_bottom {
-    padding: 0.5em;
-    display: flex;
-    flex-direction: column;
-    row-gap: 0.5em;
-}
-.reason_bottom h3 {
-    font-family: var(--font-family-grotesk);
-    font-size: 16px;
-    font-weight: 500;
-    color: var(--white);
-}
-.reason_bottom p {
-    font-family: var(--font-family-grotesk);
-    font-size: 13px;
+.section_video_cont::after {
+    content: "SmartCode";
+    position: absolute;
+    font-size: 25px;
+    background-color: var(--main-color);
     color: var(--white-smoke);
-    font-weight: 500;
+    padding: 0.5em;
+    border-radius: 10px;
+    right: 0;
+    bottom: 0px;
+    z-index: 99998;
+    font-family: var(--font-family-grotesk);
+    font-weight: 700;
+    letter-spacing: 2px;
 }
