+++ src/Components/Menu/Menu.module.css
@@ -1,6 +1,7 @@
 .header_cont {
     display: flex;
     z-index: 99999999999999;
+    height: 80px;
     background-color: transparent;
 }
 .header {
