+++ src/Pages/Home/Home.module.css
@@ -1,79 +1,4 @@
-.home_header {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.home_bg {
-    width: 100%;
-    height: 100vh;
-    position: relative;
-    transform: translateY(-80px);
-    z-index: -1;
-}
-.home_bg video {
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-}
-.home_bg::before {
-    content: "";
-    position: absolute;
-    backdrop-filter: blur(8px) grayscale(50%);
-    width: 100%;
-    height: 100%;
-}
-.header_content {
-    position: absolute;
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-    align-items: center;
-}
-.header_title {
-    font-family: var(--font-family-grotesk);
-    color: var(--main-color);
-}
-.header_text {
+#home {
     display: flex;
     flex-direction: column;
-    align-items: start;
-    row-gap: 0.5em;
-}
-.header_tagline,
-.header_intro {
-    font-size: 18px;
-    white-space: nowrap;
-    font-weight: 500;
-    color: aliceblue;
-    text-align: start;
-    font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-}
-.header_intro span {
-    color: var(--main-color);
-    font-weight: 700;
-}
-.main {
-    position: relative;
-}
-.main_bg {
-    background-image: url("../../assets//images/section.jpg");
-    background-repeat: no-repeat;
-    background-size: cover;
-    height: 90vh;
-    background-attachment: fixed;
-}
-.main_bg::before {
-    content: "";
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    left: 0;
-    backdrop-filter: blur(5px) grayscale(50%);
-}
-.main h2 {
-    position: absolute;
-    z-index: 1;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
 }
